Add revisionId to get request

+++ src/main/java/me/wieku/noteserv/database/NoteRepositoryCustom.java
@@ -2,7 +2,7 @@ package me.wieku.noteserv.database;
 
 public interface NoteRepositoryCustom {
     Note getNewestNote(long noteId);
-    Note getNoteRevision(long noteId, long revision);
+    Note getNoteRevision(long noteId, int revision);
     void addNote(Note note);
     void updateNote(long noteId, Note note);
     void removeNote(long noteId);
